requirements.txt
flask==3.1.0
requests==2.32.3
Flask-SQLAlchemy==3.1.1
dnspython==2.7.0
psycopg2==2.9.10  # driver for PostgreSQL